CP-54351 move VDI from /opt
