+++ drivers/FileSR.py
@@ -18,7 +18,7 @@
 # FileSR: local-file storage repository
 
 import SR
-import VDI
+from sm import VDI
 from sm import SRCommand
 from sm.core import util
 from sm.core import scsiutil
