+++ drivers/LVHDSR.py
@@ -26,7 +26,7 @@ import glob
 
 import SR
 from SR import deviceCheck
-import VDI
+from sm import VDI
 from sm import SRCommand
 from sm.core import util
 from sm.core import scsiutil
